Extract the net/http lowering group into a host submodule
Fifth cranelift_backend.rs split under #1254, and the third of the
direct-native i64 lowering by capability group (after host_fs and
host_crypto). A call-closure analysis seeded on the net/http family
identified 61 i64 lowering functions: TCP/UDP resolve and loopback-listen
helpers, HTTP request/response parsing and server-intrinsic lowering, and
their is_i64_*/lower_i64_* name predicates and audited-expression helpers.

Move those 61 functions into
stage1/crates/axiomc/src/cranelift_backend/host_net_http.rs.
cranelift_backend.rs drops 21,948 -> 20,914 lines and the top-file line
share falls to 0.7466.

Pure function move -- no struct relocation. Unlike the crypto slice, the
Spike{Http,Tcp,Udp}* structs and the SPIKE_TCP_*/SPIKE_UDP_* registry
statics stay in the parent because the compile-time evaluator also drives
them (evaluator.rs holds the tcp/udp registry accessors); the moved
functions reach those shared structs and registries through `use super::*`,
and the fifteen functions the evaluator calls back into
(http_get/http_serve_*/net_timeout/net_tcp_listen_loopback_once/...) resolve
through the parent's `pub(crate) use host_net_http::*` re-export, the same
pattern the host_fs and host_crypto siblings established.

I64NetResolveHost is raised to pub(crate) (private_interfaces lint) because
it appears in moved pub(crate) function signatures, mirroring the
I64StaticBindings and I64HelperSignature bumps from earlier slices.

Refs #1254.
